+++ database/migrations/2026_01_15_165742_create_orders_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
-            $table->decimal('total_price', 10, 2)->default(0);
+            $table->unsignedInteger('total_price')->default(0);
             $table->timestamps();
         });
     }
