+++ database/migrations/2026_01_15_165754_create_order_products_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('order_products', function (Blueprint $table) {
             $table->id();
-            $table->decimal('price', 10, 2)->default(0);
+            $table->unsignedInteger('price')->default(0);
             $table->unsignedInteger('quantity')->default(0);
             $table->foreignId('order_id')->constrained()->cascadeOnDelete();
             $table->foreignId('product_id')->constrained()->cascadeOnDelete();
